folder structure + removing proptypes
