+++ src/config/firebaseConfig.js
@@ -1,4 +1,3 @@
-// src/config/firebaseConfig.js
 
 import { initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
