+++ src/constants/dummyData.js
@@ -1,0 +1,53 @@
+export const projects = [
+    {
+      id: 1,
+      name: "Add Project",
+      description: "",
+      isAddProject: true,
+    },
+    {
+      id: 2,
+      name: "Netflix",
+      description: "Netflix website 7fasd5f6",
+    },
+    {
+      id: 3,
+      name: "Spotify",
+      description: "Spotify clone project",
+    },
+    {
+      id: 4,
+      name: "Amazon",
+      description: "Amazon website clone",
+    },
+    {
+      id: 5,
+      name: "Google",
+      description: "Google search clone",
+    },
+    {
+      id: 6,
+      name: "Facebook",
+      description: "Facebook clone",
+    },
+    {
+      id: 7,
+      name: "Twitter",
+      description: "Twitter clone project",
+    },
+    {
+      id: 8,
+      name: "LinkedIn",
+      description: "LinkedIn clone",
+    },
+    {
+      id: 9,
+      name: "YouTube",
+      description: "YouTube clone project",
+    },
+    {
+      id: 10,
+      name: "Instagram",
+      description: "Instagram clone project",
+    },
+  ];
