+++ src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
 import {Provider} from 'react-redux'
-import store from './store.js';
+import store from './redux/store.js';
 import {Toaster} from 'react-hot-toast';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
