+++ src/redux/store.js
@@ -1,6 +1,6 @@
 import {configureStore} from '@reduxjs/toolkit';
-import authReducer from './features/auth/authSlice';
-import projectReducer from './features/project/projectSlice';
+import authReducer from './slice/authSlice';
+import projectReducer from './slice/projectSlice';
